# Make the niceName function even faster