Tested with benchmarks when repeated many thousands of times
```
| Method  | Mean     | Error     | StdDev    | Ratio | Gen0      | Gen1     | Gen2     | Allocated | Alloc Ratio |
|-------- |---------:|----------:|----------:|------:|----------:|---------:|---------:|----------:|------------:|
| Current | 5.907 ms | 0.0446 ms | 0.0417 ms |  1.00 | 1187.5000 | 156.2500 |   7.8125 |   7.14 MB |        1.00 |
| Old     | 7.913 ms | 0.0536 ms | 0.0501 ms |  1.34 | 2968.7500 | 671.8750 | 312.5000 |  17.84 MB |        2.50 |
| This    | 4.481 ms | 0.0556 ms | 0.0520 ms |  0.76 | 1085.9375 | 156.2500 |        - |   6.53 MB |        0.91 |
```